mptcp: fix mp_fastclose key byte order
1      < . 1002:1002(0) win 32792 <mp_fastclose>

doesn't work.  The kernel doesn't like the included key because
it is sent as host-endian instead of network byte order.

kernel.key is already in network byte order, so the extra htonll
needs to be removed.

This makes a test script using fastclose work with my fastclose
mptcp-next kernel branch.

Signed-off-by: Florian Westphal <fw@strlen.de>