mptcp: Add coverage for setsockopt(SOL_IP, IP_TOS)
Since support for setsockopt(SOL_IP, IP_TOS), this packetdrill test
provides coverage for setsockopt(SOL_IP, IP_TOS) and assesses it's behavior.

There are currently two tests to check if TOS value is being set - one with
TOS as 4 and other with TOS as 7. The expected behaviour is that in both cases
the outbound packet should have TOS set as 4.

The scripts also test if the value of TOS is set for new subflows created after
setsockopt() is called.

Signed-off-by: Poorva Sonparote <psonparo@redhat.com>